pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools >= 61.0"]

[project]
authors = [
  {"name" = "Intel Corporation", "email" = "none@xyz.com"},
]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
]
dependencies = [
  "psutil",
  "pytorch_lightning==2.1.3",
  "dgl==0.9.1",
  "dgllife==0.3.2",
  "rdkit==2023.3.1",
  "numba>=0.55.0",
  "numpy>=1.21.0",
  "matplotlib",
  "munch==2.5.0",
  "lmdb==1.4.1",
  "geometric-algebra-attention>=0.3.0",
  "sympy>=1.10.1",
  "jsonargparse[signatures]>=4.13.1",
  "mp_api==0.33.3",
  "emmet-core==0.64.0",
  "pydantic==1.10.12",
  "pymatgen==2023.7.20",
  "schema>=0.7.5",
  "ase>=3.22.1",
  "matgl==0.8.5",
  "einops==0.7.0",
  "mendeleev==0.14.0",
  "e3nn==0.5.1"
]
description = "PyTorch Lightning and Deep Graph Library enabled materials science deep learning pipeline"
dynamic = ["version", "readme"]
keywords = ["performance", "portability"]
license = {file = "LICENSE.md"}
name = "matsciml"
requires-python = ">=3.8"

[project.optional-dependencies]
all = [
  "matsciml[dev,symmetry,pyg]",
]
dev = [
  "black>=22.3.0",
  "flake8",
  "flake8-pyproject",
  "flake8-black",
  "flake8-bandit",
  "pre-commit",
  "bandit",
  "pytest"
]
pyg = [
  "torch_geometric==2.3.1",
  "torch_scatter==2.1.1",
  "torch_sparse==0.6.17",
  "pybind11"
]
symmetry = [
  "rowan==1.3.0.post1",
]

[tool.setuptools.packages.find]
include = ["matsciml*"]
where = ["."]

[tool.setuptools.dynamic]
readme = {file = ["README.md"]}
version = {attr = "matsciml.__version__"}
